requirements.txt
flet>=0.24.1
firebase-admin>=6.5.0
python-firebase>=1.2
python-dateutil>=2.9.0
Pillow>=10.4.0
python-dotenv>=1.0.1
pydantic>=2.9.2
requests>=2.32.3
asyncio>=3.4.3
aiohttp>=3.10.10
